optimized wholemolecules
by this inlining, wholemolecules on a large set of atoms is ~ 20% faster
on my laptop
