+++ src/core/ActionAtomistic.h
@@ -26,6 +26,7 @@
 #include "tools/Tensor.h"
 #include "tools/Pbc.h"
 #include "tools/ForwardDecl.h"
+#include "Value.h"
 #include <vector>
 #include <map>
 
@@ -253,6 +254,23 @@ inline
 const std::vector<AtomNumber> & ActionAtomistic::getUniqueLocal() const {
   return unique_local;
 }
+
+inline
+Vector ActionAtomistic::getGlobalPosition(const std::pair<std::size_t,std::size_t>& a) const {
+  Vector pos;
+  pos[0]=xpos[a.first]->data[a.second];
+  pos[1]=ypos[a.first]->data[a.second];
+  pos[2]=zpos[a.first]->data[a.second];
+  return pos;
+}
+
+inline
+void ActionAtomistic::setGlobalPosition(const std::pair<std::size_t, std::size_t>& a, const Vector& pos ) {
+  xpos[a.first]->data[a.second]=pos[0];
+  ypos[a.first]->data[a.second]=pos[1];
+  zpos[a.first]->data[a.second]=pos[2];
+}
+
 
 }
 #endif
